refactor: move Args to test, since they are never used for the actual CLI tool
